refactor(rust): break router out into separate module tree
